requirements.txt
absl-py==2.1.0
aiosmtplib==2.0.2
altair==5.3.0
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.1
blinker==1.7.0
cachetools==5.3.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Django==4.2.6
django-ipware==5.0.1
djangorestframework==3.14.0
dnspython==2.5.0
docopt==0.6.2
dogpile.cache==1.3.3
ecdsa==0.19.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
fastapi-mail==1.4.1
ffmpeg-python==0.2.0
ffprobe-python==1.0.3
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.46.0
fsspec==2023.12.1
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/future-0.18.2-py3-none-any.whl
gast==0.5.4
gcloud==0.17.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.4
importlib-resources==6.1.1
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jose==1.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jws==0.1.3
keras==3.2.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
llvmlite==0.42.0
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
namex==0.0.7
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.0
oauth2client==3.0.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.1.1
parso==0.8.4
pbr==6.0.0
pexpect==4.9.0
Pillow==10.1.0
pipreqs==0.4.13
platformdirs==4.2.1
pooch==1.8.1
prompt_toolkit==3.0.45
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybase64==1.3.2
pycparser==2.22
pycryptodome==3.4.3
pydantic==2.4.2
pydantic-settings==2.1.0
pydantic_core==2.10.1
pydeck==0.8.1b0
pydub==0.25.1
pygame==2.5.2
Pygments==2.17.2
pymongo==3.5.1
PyMySQL==1.1.0
pyodbc==5.0.1
pyparsing==3.1.1
Pyrebase==3.0.27
Pyrebase4==4.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ipware==3.0.0
python-jose==3.3.0
python-jwt==2.0.1
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.34.0
requests==2.31.0
requests-toolbelt==0.7.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
shellingham==1.5.4
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/six-1.15.0-py2.py3-none-any.whl
sklearn==0.0.post10
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
Starlette-WTF==0.4.5
stevedore==5.2.0
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.1
torch==2.1.1
torchsummary==1.5.1
torchvision==0.16.1
tornado==6.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.8.0
tzdata==2023.3
uplink==0.9.7
uritemplate==4.1.1
urllib3==1.26.18
utils==1.0.2
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
WTForms==3.1.2
yarg==0.1.9
zipp==3.17.0
